+++ jjava-jupyter/src/test/java/org/dflib/jjava/jupyter/kernel/util/GlobResolverTest.java
@@ -23,7 +23,7 @@ import java.util.stream.StreamSupport;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
-public class GlobFinderTest {
+public class GlobResolverTest {
     private static final Configuration WIN_FS = Configuration.windows().toBuilder().setWorkingDirectory("C:/dir-a").build();
     private static final Configuration UNIX_FS = Configuration.unix().toBuilder().setWorkingDirectory("/dir-a").build();
     private static final Configuration OSX_FS = Configuration.osX().toBuilder().setWorkingDirectory("/dir-a").build();
@@ -33,7 +33,7 @@ public class GlobFinderTest {
 
     @ParameterizedTest(name = "{index} :: {1}")
     @MethodSource("data")
-    public void test(Configuration fsConfig, String glob, Set<String> files, Set<String> paths) throws Exception {
+    public void resolve(Configuration fsConfig, String glob, Set<String> paths) throws Exception {
         FileSystem fs = Jimfs.newFileSystem(fsConfig);
         List<Path> roots = StreamSupport.stream(fs.getRootDirectories().spliterator(), false)
                 .collect(Collectors.toList());
@@ -56,27 +56,14 @@ public class GlobFinderTest {
             }
         }
 
-        Set<Path> expected;
-        Set<Path> actual;
-        GlobFinder finder = new GlobFinder(fs, glob);
+        GlobResolver finder = new GlobResolver(fs, glob);
 
-        expected = files.stream()
+        Set<Path> expected = paths.stream()
                 .map(fs::getPath)
                 .map(Path::normalize)
                 .map(Path::toAbsolutePath)
                 .collect(Collectors.toSet());
-        actual = StreamSupport.stream(finder.computeMatchingFiles().spliterator(), false)
-                .map(Path::normalize)
-                .map(Path::toAbsolutePath)
-                .collect(Collectors.toSet());
-        assertEquals(expected, actual, String.format("Glob files: '%s'", glob));
-
-        expected = paths.stream()
-                .map(fs::getPath)
-                .map(Path::normalize)
-                .map(Path::toAbsolutePath)
-                .collect(Collectors.toSet());
-        actual = StreamSupport.stream(finder.computeMatchingPaths().spliterator(), false)
+        Set<Path> actual = StreamSupport.stream(finder.resolve().spliterator(), false)
                 .map(Path::normalize)
                 .map(Path::toAbsolutePath)
                 .collect(Collectors.toSet());
@@ -88,18 +75,15 @@ public class GlobFinderTest {
                 Arguments.of(
                         WIN_FS,
                         "C:/*",
-                        allFilesMapped(s -> "C:/" + s),
                         allFilesAndDirsMapped(s -> "C:/" + s)
                 ),
                 Arguments.of(
                         UNIX_FS,
                         "/*",
-                        allFilesMapped(s -> "/" + s),
                         allFilesAndDirsMapped(s -> "/" + s)
                 ),
                 Arguments.of(
                         OSX_FS, "/*",
-                        allFilesMapped(s -> "/" + s),
                         allFilesAndDirsMapped(s -> "/" + s)
                 ),
 
@@ -107,262 +91,210 @@ public class GlobFinderTest {
                 Arguments.of(
                         WIN_FS,
                         "C:/*/",
-                        allDirsFlatMapped(d -> allFilesMapped(s -> "C:/" + d + "/" + s)),
                         allFilesAndDirsMapped(s -> "C:/" + s)
                 ),
                 Arguments.of(
                         UNIX_FS,
                         "/*/",
-                        allDirsFlatMapped(d -> allFilesMapped(s -> "/" + d + "/" + s)),
                         allFilesAndDirsMapped(s -> "/" + s)),
                 Arguments.of(
                         OSX_FS,
                         "/*/",
-                        allDirsFlatMapped(d -> allFilesMapped(s -> "/" + d + "/" + s)),
                         allFilesAndDirsMapped(s -> "/" + s)
                 ),
 
                 Arguments.of(
                         WIN_FS,
                         "C:/c.txt",
-                        setOf("C:/c.txt"),
                         setOf("C:/c.txt")
                 ),
                 Arguments.of(
                         UNIX_FS,
                         "/c.txt",
-                        setOf("/c.txt"),
                         setOf("/c.txt")
                 ),
                 Arguments.of(
                         OSX_FS,
                         "/c.txt",
-                        setOf("/c.txt"),
                         setOf("/c.txt")
                 ),
 
                 Arguments.of(
                         WIN_FS,
                         "C:/*/c.txt",
-                        allDirsMapped(d -> "C:/" + d + "/c.txt"),
                         allDirsMapped(d -> "C:/" + d + "/c.txt")
                 ),
                 Arguments.of(
                         UNIX_FS,
                         "/*/c.txt",
-                        allDirsMapped(d -> "/" + d + "/c.txt"),
                         allDirsMapped(d -> "/" + d + "/c.txt")
                 ),
                 Arguments.of(
                         OSX_FS,
                         "/*/c.txt",
-                        allDirsMapped(d -> "/" + d + "/c.txt"),
                         allDirsMapped(d -> "/" + d + "/c.txt")
                 ),
 
                 Arguments.of(
                         WIN_FS,
                         "C:/dir-b/*.txt",
-                        allFilesFilterMapped(f -> f.endsWith(".txt"), f -> "C:/dir-b/" + f),
                         allFilesFilterMapped(f -> f.endsWith(".txt"), f -> "C:/dir-b/" + f)
                 ),
                 Arguments.of(
                         UNIX_FS,
                         "/dir-b/*.txt",
-                        allFilesFilterMapped(f -> f.endsWith(".txt"), f -> "/dir-b/" + f),
                         allFilesFilterMapped(f -> f.endsWith(".txt"), f -> "/dir-b/" + f)
                 ),
                 Arguments.of(
                         OSX_FS,
                         "/dir-b/*.txt",
-                        allFilesFilterMapped(f -> f.endsWith(".txt"), f -> "/dir-b/" + f),
                         allFilesFilterMapped(f -> f.endsWith(".txt"), f -> "/dir-b/" + f)
                 ),
 
                 Arguments.of(
                         WIN_FS,
                         "*.pdf",
-                        allFilesFilterMapped(f -> f.endsWith(".pdf"), f -> "./" + f),
                         allFilesFilterMapped(f -> f.endsWith(".pdf"), f -> "./" + f)
                 ),
                 Arguments.of(
                         UNIX_FS,
                         "*.pdf",
-                        allFilesFilterMapped(f -> f.endsWith(".pdf"), f -> "./" + f),
                         allFilesFilterMapped(f -> f.endsWith(".pdf"), f -> "./" + f)
                 ),
                 Arguments.of(
                         OSX_FS,
                         "*.pdf",
-                        allFilesFilterMapped(f -> f.endsWith(".pdf"), f -> "./" + f),
                         allFilesFilterMapped(f -> f.endsWith(".pdf"), f -> "./" + f)
                 ),
 
                 Arguments.of(
                         WIN_FS,
                         "*/dir.c/*.svg",
-                        allDirsFlatMapped(d -> allFilesFilterMapped(f -> f.endsWith(".svg"), f -> "./" + d + "/dir.c/" + f)),
                         allDirsFlatMapped(d -> allFilesFilterMapped(f -> f.endsWith(".svg"), f -> "./" + d + "/dir.c/" + f))
                 ),
                 Arguments.of(
                         UNIX_FS,
                         "*/dir.c/*.svg",
-                        allDirsFlatMapped(d -> allFilesFilterMapped(f -> f.endsWith(".svg"), f -> "./" + d + "/dir.c/" + f)),
                         allDirsFlatMapped(d -> allFilesFilterMapped(f -> f.endsWith(".svg"), f -> "./" + d + "/dir.c/" + f))
                 ),
                 Arguments.of(
                         OSX_FS,
                         "*/dir.c/*.svg",
-                        allDirsFlatMapped(d -> allFilesFilterMapped(f -> f.endsWith(".svg"), f -> "./" + d + "/dir.c/" + f)),
                         allDirsFlatMapped(d -> allFilesFilterMapped(f -> f.endsWith(".svg"), f -> "./" + d + "/dir.c/" + f))
                 ),
 
                 Arguments.of(
                         WIN_FS,
                         "?.pdf",
-                        setOf("C:/dir-a/a.pdf"),
                         setOf("C:/dir-a/a.pdf")
                 ),
                 Arguments.of(
                         UNIX_FS,
                         "?.pdf",
-                        setOf("/dir-a/a.pdf"),
                         setOf("/dir-a/a.pdf")
                 ),
                 Arguments.of(
                         OSX_FS,
                         "?.pdf",
-                        setOf("/dir-a/a.pdf"),
                         setOf("/dir-a/a.pdf")
                 ),
 
                 Arguments.of(
                         WIN_FS,
                         "C:/dir-?/?.pdf",
-                        setOf("C:/dir-a/a.pdf", "C:/dir-b/a.pdf"),
                         setOf("C:/dir-a/a.pdf", "C:/dir-b/a.pdf")
                 ),
                 Arguments.of(
                         UNIX_FS,
                         "/dir-?/?.pdf",
-                        setOf("/dir-a/a.pdf", "/dir-b/a.pdf"),
                         setOf("/dir-a/a.pdf", "/dir-b/a.pdf")
                 ),
                 Arguments.of(
                         OSX_FS,
                         "/dir-?/?.pdf",
-                        setOf("/dir-a/a.pdf", "/dir-b/a.pdf"),
                         setOf("/dir-a/a.pdf", "/dir-b/a.pdf")
                 ),
 
                 Arguments.of(
                         WIN_FS,
                         "C:/dir.c/abc.svg",
-                        setOf("C:/dir.c/abc.svg"),
                         setOf("C:/dir.c/abc.svg")
                 ),
                 Arguments.of(
                         UNIX_FS,
                         "/dir.c/abc.svg",
-                        setOf("/dir.c/abc.svg"),
                         setOf("/dir.c/abc.svg")
                 ),
                 Arguments.of(
                         OSX_FS,
                         "/dir.c/abc.svg",
-                        setOf("/dir.c/abc.svg"),
                         setOf("/dir.c/abc.svg")
                 ),
 
                 Arguments.of(
                         WIN_FS,
                         "C:/bad",
-                        Collections.emptySet(),
                         Collections.emptySet()
                 ),
                 Arguments.of(
                         UNIX_FS,
                         "/bad",
-                        Collections.emptySet(),
                         Collections.emptySet()
                 ),
                 Arguments.of(
                         OSX_FS,
                         "/bad",
-                        Collections.emptySet(),
                         Collections.emptySet()
                 ),
 
                 Arguments.of(
                         WIN_FS,
                         "C:/*/*/*/*/*/*/*",
-                        Collections.emptySet(),
                         Collections.emptySet()
                 ),
                 Arguments.of(
                         UNIX_FS,
                         "/*/*/*/*/*/*/*",
-                        Collections.emptySet(),
                         Collections.emptySet()
                 ),
                 Arguments.of(
                         OSX_FS,
                         "/*/*/*/*/*/*/*",
-                        Collections.emptySet(),
                         Collections.emptySet()
                 ),
 
                 Arguments.of(
                         WIN_FS,
                         "C:/dir-?/",
-                        allDirsFlatMapped(d -> !d.startsWith("dir-")
-                                ? Collections.emptySet()
-                                : allFilesMapped(f -> "C:/" + d + "/" + f)),
                         setOf("C:/dir-a", "C:/dir-b")
                 ),
                 Arguments.of(
                         UNIX_FS,
                         "/dir-?/",
-                        allDirsFlatMapped(d -> !d.startsWith("dir-")
-                                ? Collections.emptySet()
-                                : allFilesMapped(f -> "/" + d + "/" + f)),
                         setOf("/dir-a", "/dir-b")
                 ),
                 Arguments.of(
                         OSX_FS,
                         "/dir-?/",
-                        allDirsFlatMapped(d -> !d.startsWith("dir-")
-                                ? Collections.emptySet()
-                                : allFilesMapped(f -> "/" + d + "/" + f)),
                         setOf("/dir-a", "/dir-b")
                 ),
 
                 Arguments.of(
                         WIN_FS,
                         "C:/*c*",
-                        allFilesFilterMapped(f -> f.contains("c"), f -> "C:/" + f),
                         allFilesAndDirsFilterMapped(f -> f.contains("c"), f -> "C:/" + f)
                 ),
                 Arguments.of(
                         UNIX_FS,
                         "/*c*",
-                        allFilesFilterMapped(f -> f.contains("c"), f -> "/" + f),
                         allFilesAndDirsFilterMapped(f -> f.contains("c"), f -> "/" + f)
                 ),
                 Arguments.of(
                         OSX_FS,
                         "/*c*",
-                        allFilesFilterMapped(f -> f.contains("c"), f -> "/" + f),
                         allFilesAndDirsFilterMapped(f -> f.contains("c"), f -> "/" + f)
                 )
         );
-    }
-
-    private static Set<String> allFilesMapped(Function<String, String> mapper) {
-        return TEST_FILES_ALL
-                .stream()
-                .map(mapper)
-                .collect(Collectors.toSet());
     }
 
     private static Set<String> allFilesFilterMapped(Predicate<String> filter, Function<String, String> mapper) {